+++ domain/src/main/java/tachiyomi/domain/history/model/HistorySourceFilter.kt
@@ -3,4 +3,5 @@ package tachiyomi.domain.history.model
 sealed interface HistorySourceFilter {
     data class Library(val excludedLocalSourceId: Long) : HistorySourceFilter
     data class Local(val localSourceId: Long) : HistorySourceFilter
+    data class Source(val excludedLocalSourceId: Long) : HistorySourceFilter
 }